lazy load gum provider
