+++ next.config.js
@@ -1,4 +1,9 @@
 /** @type {import('next').NextConfig} */
+
+const withBundleAnalyzer = require("@next/bundle-analyzer")({
+  enabled: process.env.ANALYZE === "true",
+});
+
 const nextConfig = {
   reactStrictMode: true,
   images: {
@@ -25,4 +30,4 @@ const nextConfig = {
   transpilePackages: ["react-tweet"],
 };
 
-module.exports = nextConfig;
+module.exports = withBundleAnalyzer(nextConfig);
